merge-lock-jenkins-service/requirements.txt
python-dotenv==0.6.1
jenkinsapi==0.3.4
